// src/server/permissions/permissions.enum.ts
export enum Perms {
  CREATE = 'create',
  FIND = 'find',
  FIND_ONE = 'find_one',
  UPDATE = 'update',
  DELETE = 'delete',
  SHOW_PROFILE = 'show_profile',
  CHANGE_PASSWORD = 'change_password',
  CHANGE_EMAIL = 'change_email',
  CHANGE_USERNAME = 'change_username',
  CREATE_POSITION = 'create_position',
  CREATE_APPLICATION = 'create_application',
  CREATE_COMMENT = 'create_comment',
  ACCEPT_APPLICATION = 'accept_application',
  OPEN_APPLICATION = 'open_application',
  REJECT_APPLICATION = 'reject_application',
}